Turning the js indent.

+++ public/js/ctrl.js
@@ -31,14 +31,14 @@ App.BookController = Ember.ObjectController.extend({
 			this.set('isEditing',false);
 			this.get('model').save();
 		},
-        	confirmRemove: function(){
+        confirmRemove: function(){
 			var books = this.get('model');
 			books.deleteRecord();
 			books.save();
-            		this.set('isRemoving',false);
-        	},
-        	cancelRemove: function(){
-            		this.set('isRemoving',false);
+            this.set('isRemoving',false);
+        },
+        cancelRemove: function(){
+            this.set('isRemoving',false);
 		}
 	}
 });
